[App] Enable swiftlint, fix warnings

+++ app-ios/wifilamp/Models/Devices/Device.swift
@@ -24,6 +24,6 @@ protocol Device {
     var chipId: String { get }
     
     func setup(progressUpdate: ((_ description: String, _ stepNumber: Int, _ totalSteps: Int) -> Void)?, delegate: DeviceSetupDelegate?) -> Promise<Void>
-    func setColor(on deviceUrl: URL, _ red: CGFloat, _ blue: CGFloat, _ green: CGFloat) -> Promise<VoidResponse>
+    func setColor(on deviceUrl: URL, _ red: Int, _ blue: Int, _ green: Int) -> Promise<VoidResponse>
     func turn(on isOn: Bool, on deviceUrl: URL) -> Promise<VoidResponse>
 }
